+++ app/adms/Views/confEmails/listConfEmails.php
@@ -22,7 +22,7 @@ foreach ($this->data['listConfEmails'] as $confEmail){
     echo "E-mail: $email</span><br>";
     echo "<a href='".URLADM."view-conf-emails/index/$id'>Visualizar</a><br>";
     echo "<a href='".URLADM."edit-conf-emails/index/$id'>Editar</a><br>";
-    echo "<a href='".URLADM."delete-conf-emails/index/$id'>Apagar</a><br>";
+    echo "<a href='".URLADM."delete-conf-emails/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir o registro?\")'>Apagar</a><br>";
     echo "<hr>";
 }
